refactor(dh): move format logic to pipe
